Comments and rearrange
